requirements.txt
Flask==3.0.0
groq==0.4.2
python-dotenv==1.0.1
gunicorn==21.2.0
pypdf==4.0.1
